backend/buyer-agent/requirements.txt
httpx
stellar-sdk
python-dotenv
google-generativeai
